Split verification into method-specific classes
